Small update to Tetris_2 that fixes some bugs.
-Fixes score to properly add 25 points per line
-Fixes level to properly display the current level

+++ Tetris_2/Tracer System/Grid.h
@@ -42,7 +42,7 @@ private:
 	bool lineEliminationMode;
 
 	Block m_Grid[180];
-	int score;
+	float score;
 	int level;
 	sf::Texture img_block;
 	sf::Sprite spt_block;
